+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 # Release notes
 
-## Next
+- Remove react-bootstrap from the GenomePositionSearchBox
+- Fix color issue in GeneAnnotation SVG export
+
+## v1.11.8
 
 - Remove dependency on `cwise`
 - Make vertical chromosome labels as well as loading status labels readable in vertical tracks
 - Added an option menu item for rectangle domain fill opacity
+- Added a parameter in `zoomToGene` to allow specifying padding around gene
+- Add data fetchers to `AVAILABLE_FOR_PLUGINS`
+- Update track list in `AVAILABLE_FOR_PLUGINS`
+- Correctly setup initial scales of vertical tracks when the width of a center track is zero.
+- Config-wise, allow axis-specific location locks (e.g., lock the vertical axis in a view to the horizontal axis in another).
+- Add `reload` implementation to `HiGlassComponenet` API.
 
 _[Detailed changes since v1.11.5](https://github.com/higlass/higlass/compare/v1.11.7...develop)_